feat(estructura): reorganización del proyecto a arquitectura orientada a dominios con rutas funcionales
- Componentes del Hero movidos a /features/hero con su store, contenido y types
- Navbar movido a /shared/components/Navbar.tsx como componente reutilizable
- Actualización de todos los imports a la nueva estructura con aliases (@/features, @/shared)
- Restaurada la estructura del App Router en /app con layout.tsx y rutas activas:
  - Home (/)
  - Discover (/discover)
  - Categorías (/categories) dentro de (rutas)
- Añadido types.ts en el dominio hero para el tipado del contenido y el estado
- Eliminación de carpetas antiguas tras la migración

✅ Proyecto ahora más limpio, escalable y mantenible para crecer por funcionalidades
