--- pixela-frontend/src/app/(media)/categories/[category]/movie.tsx
@@ -1,5 +1,0 @@
-//TODO Página con las películas de la categoría
-
-export default function Movie() {
-  return <div>Movie</div>;
-}
