--- pixela-frontend/src/app/(media)/categories/[category]/page.tsx
@@ -1,5 +1,0 @@
-//TODO Página para elegir entre Películas o Series dentro de una categoría
-
-export default function Category() {
-  return <div>Category</div>;
-}
