--- pixela-frontend/src/app/(media)/categories/[category]/tv.tsx
@@ -1,5 +1,0 @@
-//TODO Página con las series de la categoría
-
-export default function Tv() {
-  return <div>Tv</div>;
-}
