--- pixela-frontend/src/app/(media)/categories/page.tsx
@@ -1,6 +1,0 @@
-//TODO Página principal que lista todas las categorías (como Thrillers, Acción, Drama)
-
-export default function Categories() {
-  return <div>Categories</div>;
-}
-
