--- pixela-frontend/src/app/(media)/discovered/[category]/page.tsx
@@ -1,6 +1,0 @@
-//TODO Página que muestra el listado completo de series o películas
-
-export default function Discovered() {
-  return <div>Discovered list</div>;
-}
-
