--- pixela-frontend/src/app/(media)/discovered/page.tsx
@@ -1,6 +1,0 @@
-//TODO Página que muestra el diseño de "Discovered" con las recomendaciones
-
-export default function Discovered() {
-  return <div>Discovered</div>;
-}
-
