--- pixela-frontend/src/app/(media)/index.ts
@@ -1,1 +1,0 @@
-
