--- pixela-frontend/src/app/(media)/movies/[movieId]/page.tsx
@@ -1,6 +1,0 @@
-//TODO Página para mostrar los detalles de una película
-
-export default function Movie() {
-  return <div>Movie</div>;
-}
-
