--- pixela-frontend/src/app/(media)/tv/[serieId]/page.tsx
@@ -1,6 +1,0 @@
-//TODO Página para mostrar los detalles de una serie
-
-export default function Serie() {
-  return <div>Serie</div>;
-}
-
