+++ pixela-frontend/src/app/(rutas)/categories/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const CategoriesPage = () => {
+  return (
+    <h1>categories page</h1>
+  )
+}
+
+export default CategoriesPage;
