+++ pixela-frontend/src/app/(rutas)/discover/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const DiscoverPage = () => {
+  return (
+    <h1>Discover Page</h1>
+  )
+}
+
+export default DiscoverPage;
