--- pixela-frontend/src/app/components/series/README.md
@@ -1,16 +1,0 @@
-# Componentes de Series
-
-Esta carpeta contiene componentes específicos relacionados con la funcionalidad de series.
-
-## Componentes actuales:
-
-- SeriesGrid: Grid de series que muestra las series en formato de tarjetas
-
-## Ejemplos de otros componentes que irían aquí:
-
-- SeriesDetail: Vista detallada de una serie
-- SeriesCard: Tarjeta individual de una serie
-- SeriesFilter: Filtros para buscar series
-- SeriesPlayer: Reproductor de videos de series
-
-Estos componentes están relacionados con la lógica de negocio específica de series.
