--- pixela-frontend/src/app/components/shared/README.md
@@ -1,14 +1,0 @@
-# Componentes Compartidos
-
-Esta carpeta contiene componentes que se comparten entre diferentes partes de la aplicación y tienen una funcionalidad específica pero reutilizable.
-
-## Ejemplos de componentes que irían aquí:
-
-- Layout: Layouts compartidos de la aplicación
-- Navbar: Barra de navegación principal
-- Footer: Pie de página común
-- ErrorBoundary: Componente para manejar errores
-- NotFound: Página 404 personalizada
-- SEO: Componentes para optimización SEO
-
-Estos componentes no son tan genéricos como los de UI, pero tampoco están atados a una sola funcionalidad de negocio.
