--- pixela-frontend/src/app/components/ui/README.md
@@ -1,15 +1,0 @@
-# Componentes UI
-
-Esta carpeta contiene componentes de interfaz de usuario genéricos y reutilizables.
-
-## Ejemplos de componentes que irían aquí:
-
-- Botones
-- Inputs
-- Cards
-- Loaders
-- Modales básicos
-- Tabs
-- Toggles
-
-Estos componentes deben ser genéricos y no tener lógica de negocio específica.
