--- pixela-frontend/src/app/ejemplos-fuentes/page.tsx
@@ -1,5 +1,0 @@
-import FontExamples from '../../examples/page';
-
-export default function EjemplosFuentesPage() {
-  return <FontExamples />;
-}
