+++ pixela-frontend/src/app/page.tsx
@@ -1,5 +1,5 @@
 import { heroData } from "@/data";
-import { HeroSection } from "./components/ui/HeroSection/HeroSection";
+import { HeroSection } from "@/features/hero/components/HeroSection";
 
 //TODO Página de inicio o layout principal
 export default function Home() {
