+++ pixela-frontend/src/features/hero/components/index.ts
