+++ pixela-frontend/src/features/hero/content.ts
@@ -1,0 +1,10 @@
+export const heroData = {
+    title: "Explora el universo",
+    accentTitle: "cinematográfico",
+    description: "Descubre, colecciona y comparte experiencias audiovisuales en una comunidad de apasionados del cine y las series.",
+    secondaryButtonText: "Descubrir más",
+    images: ["/source/images/Adolescence.avif", 
+      "/source/images/Blackmirror.avif", 
+      "/source/images/MovieDestacada_uno.avif", 
+      "/source/images/MovieDestacada_dos.avif"],
+  };
