+++ pixela-frontend/src/features/hero/type.ts
@@ -1,0 +1,28 @@
+// Contenido de la sección de contenido
+export interface HeroContent {
+    title: string;
+    accentTitle: string;
+    description: string;
+    secondaryButtonText: string;
+    ctaText: string;
+    ctaLink: string;
+    images: string[];
+  }
+  
+// Estado global manejado por Zustand
+export interface HeroState {
+    currentImageIndex: number;
+    fadeIn: boolean;
+    isPlaying: boolean;
+    progress: number;
+  
+    setCurrentImageIndex: (index: number) => void;
+    setFadeIn: (state: boolean) => void;
+    setIsPlaying: (state: boolean) => void;
+    setProgress: (progress: number | ((prev: number) => number)) => void;
+  
+    prevImage: (imagesLength: number) => void;
+    nextImage: (imagesLength: number) => void;
+    handleSlideChange: (index: number) => void;
+    resetProgress: () => void;
+  }
