+++ pixela-frontend/src/shared/components/navbar.tsx
